feat. 상세페이지 - 카운터인풋 컴포넌트 삭제
컴포넌트화까지는 필요하지 않은 것으로 판단
